[add] - alterações necessárias conforme a aula 05.
