--- src/main/java/br/com/alura/sintaxeBasicaFluxo/TestaCaractere.java
@@ -1,24 +1,0 @@
-/**
- * Project: desenvolvedor
- * Date: 09/06/2020
- * User: phmiranda
- * Descrição:
- */
-
-package br.com.alura.sintaxeBasicaFluxo;
-
-public class TestaCaractere {
-    public static void main(String[] args) {
-        char letra = 'a';
-        System.out.println(letra);
-
-        char valor = 65;
-        System.out.println(valor);
-
-        String palavra = "Alguma frase preenchida aqui !";
-        System.out.println(palavra);
-
-        palavra = palavra + 2020;
-        System.out.println(palavra);
-    }
-}
