--- src/main/java/br/com/alura/sintaxeBasicaFluxo/TestaFluxoRepeticao.java
@@ -1,16 +1,0 @@
-/**
- * Project: desenvolvedor
- * Date: 09/06/2020
- * User: phmiranda
- * Descrição:
- */
-
-package br.com.alura.sintaxeBasicaFluxo;
-
-public class TestaFluxoRepeticao {
-    public static void main(String[] args) {
-        for (int i = 0; i <= 100; i++){
-            System.out.println(i * 2);
-        }
-    }
-}
