--- src/main/java/br/com/alura/sintaxeBasicaFluxo/TestaMultiplo.java
@@ -1,16 +1,0 @@
-/**
- * Project: desenvolvedor
- * Date: 09/06/2020
- * User: phmiranda
- * Descrição:
- */
-
-package br.com.alura.sintaxeBasicaFluxo;
-
-public class TestaMultiplo {
-    public static void main(String[] args) {
-        for (int numero = 3; numero < 100; numero++){
-            System.out.println(numero);
-        }
-    }
-}
