--- src/main/java/br/com/alura/sintaxeBasicaFluxo/TestaVariavel.java
@@ -1,17 +1,0 @@
-/**
- * Project: desenvolvedor
- * Date: 09/06/2020
- * User: phmiranda
- * Descrição:
- */
-
-package br.com.alura.sintaxeBasicaFluxo;
-
-public class TestaVariavel {
-    public static void main(String[] args) {
-        System.out.println("Olá, executando novo teste.");
-
-        int idade = 27;
-        System.out.println("Olá, sua idade é: " + idade);
-    }
-}
