--- src/main/java/br/com/alura/sintaxeBasicaHeranca/Conta.java
@@ -1,90 +1,0 @@
-/**
- * Project: desenvolvedor
- * Date: 15/06/2020
- * User: phmiranda
- * Descrição:
- */
-
-package br.com.alura.sintaxeBasicaHeranca;
-
-public abstract class Conta {
-    private double saldo;
-    private int agencia;
-    private int numero;
-    private Cliente titular;
-    private static int total = 0;
-
-    public Conta(int agencia, int numero){
-        Conta.total++;
-        //System.out.println("O total de contas é " + Conta.total);
-        this.agencia = agencia;
-        this.numero = numero;
-        //this.saldo = 100;
-        //System.out.println("Estou criando uma conta " + this.numero);
-    }
-
-    // métodos específicos de utilização da conta.
-    public void depositar(double valor) {
-        this.saldo = this.saldo + valor;
-    }
-
-    public boolean sacar(double valor) {
-        if(this.saldo >= valor) {
-            this.saldo -= valor;
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-    public boolean transferir(double valor, Conta destino) {
-        if(this.sacar(valor)) {
-            destino.depositar(valor);
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-    // métodos getters e setters.
-    public double getSaldo(){
-        return this.saldo;
-    }
-
-    public int getNumero(){
-        return this.numero;
-    }
-
-    public void setNumero(int numero){
-        if(numero <= 0) {
-            System.out.println("Nao pode valor menor igual a 0");
-            return;
-        }
-        this.numero = numero;
-    }
-
-    public int getAgencia(){
-        return this.agencia;
-    }
-
-    public void setAgencia(int agencia){
-        if(agencia <= 0) {
-            System.out.println("Nao pode valor menor igual a 0");
-            return;
-        }
-        this.agencia = agencia;
-    }
-
-    public void setTitular(Cliente titular){
-        this.titular = titular;
-    }
-
-    public Cliente getTitular(){
-        return this.titular;
-    }
-
-    public static int getTotal(){
-        return Conta.total;
-    }
-
-}
