--- src/main/java/br/com/alura/sintaxeBasicaHeranca/ContaPoupanca.java
@@ -1,14 +1,0 @@
-/**
- * Project: desenvolvedor
- * Date: 15/06/2020
- * User: phmiranda
- * Descrição:
- */
-
-package br.com.alura.sintaxeBasicaHeranca;
-
-public class ContaPoupanca extends Conta {
-    public ContaPoupanca(int agencia, int numero) {
-        super(agencia, numero);
-    }
-}
