+++ src/main/java/br/com/alura/sintaxeBasicaHeranca/banco/ContaPoupanca.java
@@ -1,0 +1,19 @@
+/**
+ * Project: desenvolvedor
+ * Date: 15/06/2020
+ * User: phmiranda
+ * Descrição:
+ */
+
+package br.com.alura.sintaxeBasicaHeranca.banco;
+
+public class ContaPoupanca extends Conta {
+    public ContaPoupanca(int agencia, int numero) {
+        super(agencia, numero);
+    }
+
+    @Override
+    public void depositar(double valor) {
+        super.saldo += valor;
+    }
+}
