+++ src/main/java/br/com/alura/sintaxeBasicaHeranca/funcionario/FuncionarioDesignerGrafico.java
@@ -1,0 +1,15 @@
+/**
+ * Project: desenvolvedor
+ * Date: 12/06/2020
+ * User: phmiranda
+ * Descrição:
+ */
+
+package br.com.alura.sintaxeBasicaHeranca.funcionario;
+
+public class FuncionarioDesignerGrafico extends Funcionario {
+    public double getBonificacao() {
+        System.out.println("Bonificação: DESIGNER GRÁFICO");
+        return 200;
+    }
+}
