--- src/main/java/br/com/alura/sintaxeBasicaOO/Cliente.java
@@ -1,12 +1,0 @@
-/**
- * Project: desenvolvedor
- * Date: 09/06/2020
- * User: phmiranda
- * Descrição:
- */
-
-package br.com.alura.sintaxeBasicaOO;
-
-public class Cliente {
-
-}
