--- src/main/java/br/com/alura/sintaxeBasicaOO/TestaSaque.java
@@ -1,12 +1,0 @@
-/**
- * Project: desenvolvedor
- * Date: 09/06/2020
- * User: phmiranda
- * Descrição:
- */
-
-package br.com.alura.sintaxeBasicaOO;
-
-public class TestaSaque {
-
-}
