+++ src/main/java/br/com/alura/sintaxeFluxoRepeticao/TestaCaractere.java
@@ -1,0 +1,24 @@
+/**
+ * Project: desenvolvedor
+ * Date: 09/06/2020
+ * User: phmiranda
+ * Descrição:
+ */
+
+package br.com.alura.sintaxeFluxoRepeticao;
+
+public class TestaCaractere {
+    public static void main(String[] args) {
+        char letra = 'a';
+        System.out.println(letra);
+
+        char valor = 65;
+        System.out.println(valor);
+
+        String palavra = "Alguma frase preenchida aqui !";
+        System.out.println(palavra);
+
+        palavra = palavra + 2020;
+        System.out.println(palavra);
+    }
+}
