+++ src/main/java/br/com/alura/sintaxeFluxoRepeticao/TestaFluxoRepeticao.java
@@ -1,0 +1,16 @@
+/**
+ * Project: desenvolvedor
+ * Date: 09/06/2020
+ * User: phmiranda
+ * Descrição:
+ */
+
+package br.com.alura.sintaxeFluxoRepeticao;
+
+public class TestaFluxoRepeticao {
+    public static void main(String[] args) {
+        for (int i = 0; i <= 100; i++){
+            System.out.println(i * 2);
+        }
+    }
+}
