+++ src/main/java/br/com/alura/sintaxeOO/TestaMetodo.java
@@ -1,0 +1,12 @@
+/**
+ * Project: desenvolvedor
+ * Date: 09/06/2020
+ * User: phmiranda
+ * Descrição:
+ */
+
+package br.com.alura.sintaxeOO;
+
+public class TestaMetodo {
+
+}
